KINO-17 Entities lavet, så de matcher nyt ER diagram uden ekstra samlingstabel
