+++ src/main/java/org/example/kinobackend/model/BookedSeatID.java
@@ -6,7 +6,7 @@ package org.example.kinobackend.model;
 import java.io.Serializable;
 
 
-public class ShowSeatID implements Serializable {
+public class BookedSeatID implements Serializable {
     //implementering af Serializable gør det muligt at bruge nøglen som identifikator i Hibernate/JPA
     private int show;
     private int seat;
