+++ src/main/java/org/example/kinobackend/model/Status.java
@@ -2,6 +2,5 @@ package org.example.kinobackend.model;
 
 public enum Status {
     AVAILABLE,
-    BOOKED,
-    BLOCKED
+    BOOKED
 }
